[FIX]: fix #4244 [FIX]: fix #4243

+++ AgoraSyncManager/SyncManager/RtmManager/RtmSyncManager+Handle.swift
@@ -9,11 +9,12 @@ import Foundation
 import AgoraRtmKit
 
 /// 哪些行为会有回调事件？
-/// 1. sceneRef.delete() 没有回调事件
-/// 2. AgoraSyncManager.deleteScenes["id"] 删除房间列表的一个房间 监听subscribe key == nil 有onDeleted事件
-/// 3. sceneRef.update(key) 更新房间信息 只有update事件
+/// 1. sceneRef.delete() subscribe(key: nil)
+/// 3. sceneRef.update(key) subscribe(key: key) 更新房间信息 只有update事件
 /// 4. syncRef.collection(className: "member").add(data: ["userName" : "UserName"])
 /// 监听syncRef.collection(className: "member").document().subscribe(key: nil) 有齐全的事件
+///
+/// 
 
 extension RtmSyncManager {
     func notifyObserver(channel: AgoraRtmChannel, attributes: [AgoraRtmChannelAttribute]) {
@@ -38,7 +39,8 @@ extension RtmSyncManager {
                             attributes: attributes,
                             onCreateBlock: onCreateBlock,
                             onUpdateBlock: onUpdateBlock,
-                            onDeleteBlock: onDeleteBlock)
+                            onDeleteBlock: onDeleteBlock,
+                            isDefaultChannelhandle: true)
                 cachedAttrs[channel] = attributes
                 return
             }
@@ -90,11 +92,18 @@ extension RtmSyncManager {
         Log.info(text: "--- notifyObserver do empty ---", tag: "RtmSyncManager")
     }
     
+    
+    /// 计算应该产生的回调事件，并执行回调事件
+    /// - Parameters:
+    ///   - cache: 原来的数据
+    ///   - attributes: 更新的attributes
+    ///   - isDefaultChannelhandle: 是否是默认channel的处理，主要处理删除房间时，本地不产生onDelete回调事件
     fileprivate func invokeEvent(cache: [AgoraRtmChannelAttribute],
                                  attributes: [AgoraRtmChannelAttribute],
                                  onCreateBlock: OnSubscribeBlock?,
                                  onUpdateBlock: OnSubscribeBlock?,
-                                 onDeleteBlock: OnSubscribeBlock?) {
+                                 onDeleteBlock: OnSubscribeBlock?,
+                                 isDefaultChannelhandle: Bool = false) {
         
         Log.info(text: "--- onCreateBlock is \(onCreateBlock == nil ? "nil" : "not nil")", tag: "RtmSyncManager")
         Log.info(text: "--- onUpdateBlock is \(onUpdateBlock == nil ? "nil" : "not nil")", tag: "RtmSyncManager")
@@ -141,8 +150,16 @@ extension RtmSyncManager {
         
         if let onDeleteBlockTemp = onDeleteBlock {
             for i in onlyA {
-                Log.info(text: "--- invoke onDeleteBlock", tag: "RtmSyncManager")
-                onDeleteBlockTemp(i)
+                if isDefaultChannelhandle {
+                    Log.info(text: "--- invoke onDeleteBlock", tag: "RtmSyncManager")
+                    onDeleteBlockTemp(i)
+                    continue
+                }
+                
+                if isDefaultChannelhandle, i.getId() != self.sceneName { /** 过滤非本房间的删除事件 **/
+                    Log.info(text: "--- be ignore, not local scene id", tag: "RtmSyncManager")
+                    continue
+                }
             }
         }
     }
